automatically backed up by learn

+++ lib/bachelor.rb
@@ -45,7 +45,7 @@ def get_average_age_for_season(data, season)
   number_of_contestants = 0
   data[season].each do |person, person_data|
     number_of_contestants += 1
-    age_total += person["age"].to_i
+    age_total += person["age"].to_f
   end
   age_total/number_of_contestants
 end
